+++ src/utils/exorciseNode.ts
@@ -1,12 +1,6 @@
-/*
- * This is an adaptation of Alok Menghrajani's HTML sanitizer, by sidiousvic.
- https://www.quaxio.com/html_white_listed_sanitizer/
- * Takes a potentiallyDangerousHTML string, returns a sanitized (exorcised) node.
- */
+import { allowedTags, allowedCSS } from "./exorciserAllowlist";
 
-import { allowedTags, allowedCSS } from "./allowed";
-
-function exorciseNode(node: HTMLElement) {
+export default function exorciseNode(node: HTMLElement) {
   const doc = document.implementation.createHTMLDocument();
 
   const nodeName = node.nodeName.toLowerCase();
@@ -53,13 +47,4 @@ function exorciseNode(node: HTMLElement) {
     exorcisedNode.appendChild(exorciseNode(child as HTMLElement));
   }
   return exorcisedNode;
-}
-
-export default function phantomExorciser(potentiallyDangerousHTML: string) {
-  const doc = document.implementation.createHTMLDocument();
-
-  const div = doc.createElement("div");
-  div.innerHTML = potentiallyDangerousHTML;
-
-  return (exorciseNode(div) as HTMLElement).innerHTML;
 }
